requirements.txt
duckdb>=1.0
dlt[duckdb]>=1.0
pandas>=2.0
pyyaml>=6.0
requests>=2.31
python-dotenv>=1.0
pytest>=8.0
ruff>=0.6
